+++ scripts/constants.js
@@ -526,6 +526,16 @@ export const DEFAULT_SETTINGS = {
     default: true,
   },
 
+  systemConditionOverrides: {
+    name: 'PF2E_VISIONER.SETTINGS.SYSTEM_CONDITION_OVERRIDES.name',
+    hint: 'PF2E_VISIONER.SETTINGS.SYSTEM_CONDITION_OVERRIDES.hint',
+    scope: 'world',
+    config: true,
+    restricted: true,
+    type: Boolean,
+    default: false,
+  },
+
   enableHoverTooltips: {
     name: 'PF2E_VISIONER.SETTINGS.ENABLE_HOVER_TOOLTIPS.name',
     hint: 'PF2E_VISIONER.SETTINGS.ENABLE_HOVER_TOOLTIPS.hint',